Move types to new package
