+++ sdk/cloudmachine/Azure.CloudMachine/samples/README.md
@@ -1,0 +1,12 @@
+---
+page_type: sample
+languages:
+- csharp
+products:
+# Including relevant stubs from https://review.docs.microsoft.com/help/contribute/metadata-taxonomies#product
+- azure
+name: Azure.CloudMachine samples for .NET
+description: Samples for the Azure.CloudMachine client library.
+---
+
+# Azure.CloudMachine Samples
