+++ sdk/cloudmachine/Azure.CloudMachine/src/ClientCache.cs
@@ -9,9 +9,9 @@ namespace Azure.Core;
 // TODO: this is a very demo implementation. We need to do better
 public class ClientCache
 {
-    private readonly Dictionary<(Type, string?), object> _clients = new Dictionary<(Type, string?), object>();
+    private readonly Dictionary<(Type, string), object> _clients = new Dictionary<(Type, string), object>();
 
-    public T Get<T>(Func<T> value, string? id = default) where T: class
+    public T Get<T>(Func<T> value, string id = default) where T: class
     {
         var client = (typeof(T), id);
         lock (_clients)
