+++ sdk/cloudmachine/Azure.CloudMachine/src/CloudMachineWorkspace.cs
