+++ sdk/cloudmachine/Azure.CloudMachine/src/extensions/KeyVaultExtensions.cs
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using Azure.Core;
+using Azure.Security.KeyVault.Secrets;
+
+namespace Azure.CloudMachine.KeyVault;
+
+public static class KeyVaultExtensions
+{
+    public static SecretClient GetKeyVaultSecretsClient(this ClientWorkspace workspace)
+    {
+        ClientConnectionOptions connection = workspace.GetConnectionOptions(typeof(SecretClient));
+        if (connection.ConnectionKind == ClientConnectionKind.EntraId)
+        {
+            return new(connection.Endpoint, connection.TokenCredential);
+        }
+        throw new Exception("API key not supported");
+    }
+}
