+++ sdk/cloudmachine/ci.yml
@@ -1,0 +1,53 @@
+# NOTE: Please refer to https://aka.ms/azsdk/engsys/ci-yaml before editing this file.
+
+trigger:
+  branches:
+    include:
+    - main
+    - hotfix/*
+    - release/*
+  paths:
+    include:
+    - sdk/cloudmachine/
+#if (false)
+    # The following paths should only be included in template/ci.yml, and removed from any other SDKs which copy this file.
+    # The surrounding conditions should accomplish that when installed with `dotnet new azsdk`.
+    # eng/common code changes trigger template pipeline for basic checking.
+    - eng/common/
+#endif
+
+pr:
+  branches:
+    include:
+    - main
+    - feature/*
+    - hotfix/*
+    - release/*
+  paths:
+    include:
+    - sdk/cloudmachine/
+#if (false)
+    # The following paths should only be included in template/ci.yml, and removed from any other SDKs which copy this file.
+    # The surrounding conditions should accomplish that when installed with `dotnet new azsdk`.
+    # eng/common code changes trigger template pipeline for basic checking.
+    - eng/common/
+#endif
+
+# Switch to canary to test canary 1es branch. 1es template validation will set this parameter
+# to canary on run.
+parameters:
+- name: oneESTemplateTag
+  type: string
+  default: release
+  values:
+    - release
+
+extends:
+  template: /eng/pipelines/templates/stages/archetype-sdk-client.yml
+  parameters:
+    ServiceDirectory: cloudmachine
+    oneESTemplateTag: ${{ parameters.oneESTemplateTag }}
+    ArtifactName: packages
+    Artifacts:
+    - name: Azure.CloudMachine
+      safeName: AzureCloudMachine
