+++ sdk/provisioning/Azure.Provisioning.CloudMachine/src/AzureSdkExtensions/KeyVaultFeature.cs
@@ -1,12 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using Azure.Core;
 using Azure.Provisioning.Authorization;
 using Azure.Provisioning.Expressions;
 using Azure.Provisioning.KeyVault;
-using Azure.Security.KeyVault.Secrets;
 
 namespace Azure.Provisioning.CloudMachine.KeyVault;
 
@@ -59,18 +56,5 @@ public class KeyVaultFeature : CloudMachineFeature
         kvMiRoleAssignment.RoleDefinitionId = BicepFunction.GetSubscriptionResourceId("Microsoft.Authorization/roleDefinitions", KeyVaultBuiltInRole.KeyVaultAdministrator.ToString());
         kvMiRoleAssignment.PrincipalId = infrastructure.Identity.PrincipalId;
         infrastructure.AddResource(kvMiRoleAssignment);
-    }
-}
-
-public static class KeyVaultExtensions
-{
-    public static SecretClient GetKeyVaultSecretsClient(this ClientWorkspace workspace)
-    {
-        ClientConnectionOptions connection = workspace.GetConnectionOptions(typeof(SecretClient));
-        if (connection.ConnectionKind == ClientConnectionKind.EntraId)
-        {
-            return new(connection.Endpoint, connection.TokenCredential);
-        }
-        throw new Exception("API key not supported");
     }
 }
